[Refactor#54] FileManager 파일을 util 모듈로 이동
