+++ core/util/build.gradle.kts
@@ -34,6 +34,7 @@ android {
 
 dependencies {
 
+    implementation(libs.androidx.exifinterface)
     implementation(libs.androidx.core.ktx)
     implementation(libs.androidx.appcompat)
     implementation(libs.material)
